[concurrency] Extract out ExecutorTrackingInfo from Actor.cpp into TaskTrackingPrivate.h.
I am going to be using this to implement some functionality in Task.cpp... so it
makes sense to extract ExecutorTrackingInfo into its own private
TaskTrackingPrivate.h local impl header.
